backend/requirements.txt
firebase_functions~=0.1.0
firebase-admin>=6.0.0
twilio>=8.0.0
schedule>=1.2.0
SQLAlchemy>=2.0.0
pydantic>=2.0.0
email-validator>=2.0.0
python-dotenv>=1.0.0